requirements.txt
discord.py
virtualcrypto
python-dotenv
